requirements.txt
discord.py
pytz
google-api-python-client
google-auth

